updated user controller

+++ src/main/java/com/example/hbs/service/UserService.java
@@ -3,6 +3,7 @@ package com.example.hbs.service;
 import com.example.hbs.model.User;
 import com.example.hbs.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,17 +14,25 @@ public class UserService {
     @Autowired
     private UserRepository userRepository;
 
-    public void addUser(User user) {
+    public ResponseEntity<String> addUser(User user) {
         userRepository.save(user);
+        return ResponseEntity.ok("User Added");
     }
 
     public List<User> findAllUser() {
         return userRepository.findAll();
     }
 
-    public void updateUser(Long id, User user) {
-        User temp = userRepository.findById(id).get();
-        userRepository.delete(temp);
-        userRepository.save(user);
+    public ResponseEntity<String> updateUser(Long id, User user) {
+        if(userRepository.findById(id).isPresent()) {
+            User temp = userRepository.findById(id).get();
+            userRepository.delete(temp);
+            userRepository.save(user);
+            return ResponseEntity.ok("User updated");
+        }
+        else
+        {
+            return ResponseEntity.badRequest().body("No such id found");
+        }
     }
 }
